Harden invoice drafts: draftKind discriminator, split modules, unified UI.
- Require draftKind (enrollment_merge | customized_manual); reject mixed payloads.
- Use _first_present for numeric-safe field reads; ROUND_HALF_UP tax quantize; reject
  taxAmount + taxRate together; reject description > 500 chars; richer customized audit.
- Split admin_billing_invoice_drafts.py + admin_billing_invoice_draft_helpers.py (<500 lines).
- Migration downgrade deletes invoices with only null-enrollment lines.
- OpenAPI: oneOf + discriminator, enrollmentId oneOf uuid|null, inventory line restored.
- Admin: single Create draft invoice card with mode switch; CustomizedDraftInvoiceCard;
  lazy parent pickers; merged createDraftInvoice; formatTruncatedId accepts null.
- Tests: billing pytest additions; vitest updates + customized card unit test.
